backend/requirements.txt
# Web framework and server
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.4.2

# Data science libraries
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0

# Data handling and processing
joblib>=1.3.0
python-dotenv>=1.0.0
openpyxl>=3.1.0

# AI and code execution
google-genai>=1.0.0

# Database connectivity (NEW - for SQL integration)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
pymysql>=1.1.0  # MySQL adapter

# Vanna AI for enhanced text-to-SQL
vanna>=0.5.0
chromadb>=0.4.0  # Vector store for Vanna


# Optional: Additional database adapters
# redshift-connector>=2.0.0  # Amazon Redshift
# snowflake-connector-python>=3.0.0  # Snowflake
# google-cloud-bigquery>=3.0.0  # BigQuery
# cx_Oracle>=8.0.0  # Oracle

# Optional: for more advanced ML capabilities
# xgboost>=2.0.0
# lightgbm>=4.0.0

# Optional: for deep learning
# tensorflow>=2.14.0
# torch>=2.1.0
